In this project VStack, Image, Alert,Gradient, Button
Day 20, Day 21 Day 22

Challenge
One of the best ways to learn is to write your own code as often as possible, so here are three ways you should try extending this app to make sure you fully understand what’s going on:

Add an @State property to store the user’s score, modify it when they get an answer right or wrong, then display it in the alert.
Show the player’s current score in a label directly below the flags.
When someone chooses the wrong flag, tell them their mistake in your alert message – something like “Wrong! That’s the flag of France,” for example.